27oct: add pundit - products policy

+++ app/controllers/products_controller.rb
@@ -1,29 +1,33 @@
 class ProductsController < ApplicationController
   before_action :set_product, only: [:show, :edit, :update, :destroy]
-  before_action :authenticate_admin!, except: [:index, :show]
+  # before_action :authenticate_admin!, except: [:index, :show]
+  before_action :authenticate_user!, except: [:index, :show]
 
   # GET /products
   def index
     @products = Product.all
+    # authorize @products
   end
-
+  
   # GET /products/1
   def show
   end
-
+  
   # GET /products/new
   def new
     @product = Product.new
+    authorize @product
   end
-
+  
   # GET /products/1/edit
   def edit
+    authorize @product
   end
-
+  
   # POST /products
   def create
     @product = Product.new(product_params)
-
+    authorize @product
     if @product.save
       redirect_to @product, notice: 'Product was successfully created.'
     else
@@ -33,6 +37,7 @@ class ProductsController < ApplicationController
 
   # PATCH/PUT /products/1
   def update
+    authorize @product
     if @product.update(product_params)
       redirect_to @product, notice: 'Product was successfully updated.'
     else
@@ -42,6 +47,7 @@ class ProductsController < ApplicationController
 
   # DELETE /products/1
   def destroy
+    authorize @product
     if @product.destroy
       redirect_to products_url, notice: 'Product was successfully deleted.'
     else
